ganeti-webmgr/requirements.txt
Twisted>=11.0.0
django-fields
django-haystack==1.2.1
django-include-strip-tag
django-registration>=0.7,<0.8
django>=1.4,<1.5
django_object_log==0.7
django_object_permissions==1.4.3
pyopenssl
requests>=0.9
simplejson
south>=0.7.3
whoosh>=1.8.1
